# node01/hyper_server/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Hyperledger Explorer REST API Swagger",
        "description": "Rest API ",
        "contact": {
            "name": "paypro Team"
        },
        "version": "1.0.0",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3010"
        },
        {
            "url": "https://localhost:3010"
        }
    ],
    "schemes": [
        "http"
    ],
...
    "paths": {
...
        "/api/query/insertUserTx": {
            "post": {
                "tags": [
                    "chaincode"
                ],
                "summary": "invoke user Tx",
                "operationId": "insertUserTx",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Status values that need to be considered for filter",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/insertUserTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    }
                }
            }
        },
        "/api/query/selectAllUser": {
            "post": {
                "tags": [
                    "chaincode"
                ],
                "summary": "invoke user Tx",
                "operationId": "selectAllUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Status values that need to be considered for filter",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/selectAllUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    }
                }
            }
        },
        "/api/query/selectUserTx": {
            "post": {
                "tags": [
                    "chaincode"
                ],
                "summary": "invoke user Tx",
                "operationId": "selectUserTx",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Status values that need to be considered for filter",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/selectUserTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "insertUserTx": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example":"1"
                },
                "user_id": {
                    "type": "string",
                    "example":"example@paypro.co.kr"
                },
                "merchant": {
                    "type": "string",
                    "example":"삼성"
                },
                "symbol": {
                    "type": "string",
                    "example":"h3c"
                },
                "quantity": {
                    "type": "integer",
                    "example":100
                }
            }
        },
        "selectAllUser": {
            "type": "object",
            "properties": {
                "symbol": {
                    "type": "string"
                }
            }
        },
        "selectUserTx": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
    }
}